+++ log.h
@@ -5,6 +5,7 @@
 
 #include "common.h"
 
+
 using namespace std;
